back/requirements.txt
amqp==5.0.6
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.16.63
botocore==1.19.63
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cssutils==2.3.0
defusedxml==0.7.0rc2
Django==3.1.5
django-allauth==0.44.0
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-countries==7.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-js-asset==1.2.2
django-modeltranslation==0.16.2
django-mptt==0.11.0
django-rosetta==0.9.6
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
polib==1.1.1
prompt-toolkit==3.0.18
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycparser==2.20
PyJWT==1.7.1
pynliner==0.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
soupsieve==2.2.1
sqlparse==0.4.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
